restructuring files in directory
