--- readnext/evaluation/__init__.py
@@ -1,33 +1,0 @@
-from readnext.evaluation.metrics import (
-    average_precision,
-    cosine_similarity,
-    cosine_similarity_from_df,
-    count_common_citations_from_df,
-    count_common_references_from_df,
-    mean_average_precision,
-)
-from readnext.evaluation.model_scorer import (
-    CitationModelScorer,
-    FeatureWeights,
-    LanguageModelScorer,
-)
-from readnext.evaluation.pairwise_scores import (
-    precompute_co_citations,
-    precompute_co_references,
-    precompute_cosine_similarities,
-)
-
-__all__ = [
-    "average_precision",
-    "cosine_similarity",
-    "cosine_similarity_from_df",
-    "count_common_citations_from_df",
-    "count_common_references_from_df",
-    "mean_average_precision",
-    "CitationModelScorer",
-    "FeatureWeights",
-    "LanguageModelScorer",
-    "precompute_co_citations",
-    "precompute_co_references",
-    "precompute_cosine_similarities",
-]
