+++ readnext/evaluation/scoring/__init__.py
@@ -1,0 +1,40 @@
+from readnext.evaluation.scoring.hybrid_score import HybridScore, compare_hybrid_scores
+from readnext.evaluation.scoring.hybrid_scorer import HybridScorer
+from readnext.evaluation.scoring.metrics import (
+    PairwiseMetric,
+    average_precision,
+    cosine_similarity,
+    cosine_similarity_from_df,
+    count_common_citations_from_df,
+    count_common_references_from_df,
+    mean_average_precision,
+)
+from readnext.evaluation.scoring.model_scorer import (
+    CitationModelScorer,
+    FeatureWeights,
+    LanguageModelScorer,
+)
+from readnext.evaluation.scoring.precompute_scores import (
+    precompute_co_citations,
+    precompute_co_references,
+    precompute_cosine_similarities,
+)
+
+__all__ = [
+    "HybridScore",
+    "compare_hybrid_scores",
+    "HybridScorer",
+    "PairwiseMetric",
+    "average_precision",
+    "cosine_similarity",
+    "cosine_similarity_from_df",
+    "count_common_citations_from_df",
+    "count_common_references_from_df",
+    "mean_average_precision",
+    "CitationModelScorer",
+    "FeatureWeights",
+    "LanguageModelScorer",
+    "precompute_co_citations",
+    "precompute_co_references",
+    "precompute_cosine_similarities",
+]
