+++ readnext/evaluation/scoring/hybrid_score.py
@@ -1,0 +1,59 @@
+from dataclasses import dataclass
+
+import pandas as pd
+from typing_extensions import Self
+
+from readnext.evaluation.scoring.hybrid_scorer import HybridScorer
+
+
+@dataclass
+class HybridScore:
+    language_model_name: str
+    citation_to_language: float
+    citation_to_language_candidates: float
+    language_to_citation: float
+    language_to_citation_candidates: float
+
+    @classmethod
+    def from_scorer(cls, hybrid_scorer: HybridScorer) -> Self:
+        return cls(
+            language_model_name=hybrid_scorer.language_model_name,
+            citation_to_language=hybrid_scorer.citation_to_language_scores,
+            citation_to_language_candidates=hybrid_scorer.citation_to_language_candidate_scores,
+            language_to_citation=hybrid_scorer.language_to_citation_scores,
+            language_to_citation_candidates=hybrid_scorer.language_to_citation_candidate_scores,
+        )
+
+    def __str__(self) -> str:
+        return (
+            f"Language Model: {self.language_model_name}\n"
+            "---------------------\n"
+            f"Citation -> Language\n"
+            "---------------------\n"
+            f"Candidates Score: {self.citation_to_language_candidates:.3f}\n"
+            f"Final Score: {self.citation_to_language:.3f}\n\n"
+            f"Language -> Citation\n"
+            "---------------------\n"
+            f"Candidates Score: {self.language_to_citation_candidates:.3f}\n"
+            f"Final Score: {self.language_to_citation:.3f}\n"
+        )
+
+    def to_frame(self) -> pd.DataFrame:
+        return pd.DataFrame(
+            {
+                "Language Model": self.language_model_name,
+                "Citation -> Language Candidates": round(
+                    self.citation_to_language_candidates, ndigits=3
+                ),
+                "Citation -> Language Final": round(self.citation_to_language, ndigits=3),
+                "Language -> Citation Candidates": round(
+                    self.language_to_citation_candidates, ndigits=3
+                ),
+                "Language -> Citation Final": round(self.language_to_citation, ndigits=3),
+            },
+            index=[0],
+        )
+
+
+def compare_hybrid_scores(*hybrid_scores: HybridScore) -> pd.DataFrame:
+    return pd.concat([hybrid_score.to_frame() for hybrid_score in hybrid_scores], ignore_index=True)
