+++ readnext/evaluation/scoring/hybrid_scorer.py
@@ -1,17 +1,17 @@
 from dataclasses import dataclass, field
 
 import pandas as pd
-from typing_extensions import Self
 
-from readnext.evaluation import CitationModelScorer, FeatureWeights, LanguageModelScorer
-from readnext.modeling import (
-    CitationModelData,
-    LanguageModelData,
+from readnext.evaluation.scoring.model_scorer import (
+    CitationModelScorer,
+    FeatureWeights,
+    LanguageModelScorer,
 )
+from readnext.modeling import CitationModelData, LanguageModelData
 
 
 @dataclass(kw_only=True)
-class HybridRecommender:
+class HybridScorer:
     language_model_name: str
     language_model_data: LanguageModelData
     citation_model_data: CitationModelData
@@ -118,56 +118,3 @@ class HybridRecommender:
         self.score_citation_to_language(feature_weights, n_candidates, n_final)
         self.top_n_language_to_citation(feature_weights, n_candidates, n_final)
         self.score_language_to_citation(feature_weights, n_candidates, n_final)
-
-
-@dataclass
-class HybridScore:
-    language_model_name: str
-    citation_to_language: float
-    citation_to_language_candidates: float
-    language_to_citation: float
-    language_to_citation_candidates: float
-
-    @classmethod
-    def from_recommender(cls, hybrid_recommender: HybridRecommender) -> Self:
-        return cls(
-            language_model_name=hybrid_recommender.language_model_name,
-            citation_to_language=hybrid_recommender.citation_to_language_scores,
-            citation_to_language_candidates=hybrid_recommender.citation_to_language_candidate_scores,
-            language_to_citation=hybrid_recommender.language_to_citation_scores,
-            language_to_citation_candidates=hybrid_recommender.language_to_citation_candidate_scores,
-        )
-
-    def __str__(self) -> str:
-        return (
-            f"Language Model: {self.language_model_name}\n"
-            "---------------------\n"
-            f"Citation -> Language\n"
-            "---------------------\n"
-            f"Candidates Score: {self.citation_to_language_candidates:.3f}\n"
-            f"Final Score: {self.citation_to_language:.3f}\n\n"
-            f"Language -> Citation\n"
-            "---------------------\n"
-            f"Candidates Score: {self.language_to_citation_candidates:.3f}\n"
-            f"Final Score: {self.language_to_citation:.3f}\n"
-        )
-
-    def to_frame(self) -> pd.DataFrame:
-        return pd.DataFrame(
-            {
-                "Language Model": self.language_model_name,
-                "Citation -> Language Candidates": round(
-                    self.citation_to_language_candidates, ndigits=3
-                ),
-                "Citation -> Language Final": round(self.citation_to_language, ndigits=3),
-                "Language -> Citation Candidates": round(
-                    self.language_to_citation_candidates, ndigits=3
-                ),
-                "Language -> Citation Final": round(self.language_to_citation, ndigits=3),
-            },
-            index=[0],
-        )
-
-
-def compare_hybrid_scores(*hybrid_scores: HybridScore) -> pd.DataFrame:
-    return pd.concat([hybrid_score.to_frame() for hybrid_score in hybrid_scores], ignore_index=True)
