--- readnext/modeling/hybrid_models/__init__.py
@@ -1,7 +1,0 @@
-from readnext.modeling.hybrid_models.hybrid_recommender import (
-    HybridRecommender,
-    HybridScore,
-    compare_hybrid_scores,
-)
-
-__all__ = ["HybridRecommender", "HybridScore", "compare_hybrid_scores"]
